Add random seed Param to Scala CrossValidator

+++ mllib/src/main/scala/org/apache/spark/mllib/util/MLUtils.scala
@@ -264,7 +264,7 @@ object MLUtils {
    * element, the validation data, containing a unique 1/kth of the data. Where k=numFolds.
    */
   @Since("1.0.0")
-  def kFold[T: ClassTag](rdd: RDD[T], numFolds: Int, seed: Int): Array[(RDD[T], RDD[T])] = {
+  def kFold[T: ClassTag](rdd: RDD[T], numFolds: Int, seed: Long): Array[(RDD[T], RDD[T])] = {
     val numFoldsF = numFolds.toFloat
     (1 to numFolds).map { fold =>
       val sampler = new BernoulliCellSampler[T]((fold - 1) / numFoldsF, fold / numFoldsF,
